# public/openapi.yaml
openapi: 3.1.0
info:
  title: BegaIA API
  version: 0.1.0
  description: |
    API pública mínima para el asistente hotelero. Este documento resume los endpoints validados por la suite core.
servers:
  - url: /
paths:
  /api/chat:
    post:
      summary: Enviar mensaje al chat
      description: Devuelve un ACK estable; reintentos con el mismo messageId responden con deduped=true.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hotelId:
                  type: string
                  example: hotel999
                channel:
                  type: string
                  example: web
                conversationId:
                  type: string
                guestId:
                  type: string
                messageId:
                  type: string
                  description: Idempotencia por este ID provisto por el cliente
                  example: msg-123
                content:
                  type: string
                  example: "Hola, ¿pueden corroborar mi reserva?"
                mode:
                  type: string
                  example: default
              required: [content]
            examples:
              basico:
                summary: Envío básico web
                value:
                  channel: web
                  conversationId: conv-1
                  content: "Hola"
                  messageId: msg-1
              idempotente:
                summary: Reintento con el mismo messageId
                value:
                  channel: web
                  conversationId: conv-1
                  content: "Hola"
                  messageId: msg-1
      responses:
        "200":
          description: ACK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatAck"
              examples:
                ok:
                  summary: Respuesta con ACK inicial
                  value:
                    conversationId: conv-1
                    status: sent
                    message:
                      hotelId: hotel999
                      conversationId: conv-1
                      channel: web
                      messageId: msg-1
                      status: sent
                    lang: es
                dedupe:
                  summary: ACK en replay idempotente
                  value:
                    conversationId: conv-1
                    status: sent
                    message:
                      hotelId: hotel999
                      conversationId: conv-1
                      channel: web
                      messageId: msg-1
                      status: sent
                    lang: es
                    deduped: true
...
components:
  schemas:
    ChatAck:
      type: object
      properties:
        conversationId: { type: string }
        status: { type: string, enum: [sent, pending] }
        message:
          type: object
          properties:
            hotelId: { type: string }
            conversationId: { type: string }
            channel: { type: string }
            messageId: { type: string }
            status: { type: string }
            suggestion:
              type: object
              nullable: true
              additionalProperties: true
        lang: { type: string }
        deduped: { type: boolean }
...